refactor(jina): 게시글과 관련된 ui를 /entities로 이동 #51
