+++ src/entities/post/ui/CreateForm/styled.ts
