+++ src/entities/post/ui/DeleteModal/index.tsx
