--- src/features/musicRecommendationBoard/ui/MusicRecommendationForm/index.tsx
@@ -1,1 +1,0 @@
-export { MusicRecommendationForm } from './MusicRecommendationForm';
