+++ src/pages/createPost/ui/styled.ts
