+++ src/shared/ui/Input/SearchInputWithFilter.tsx
@@ -50,7 +50,7 @@ const SearchFilter = ({
   const [selectedOption, setSelectedOption] = useState<Option | null>(null);
   const ref = useRef<HTMLUListElement>(null);
   const selectRef = useRef<HTMLDivElement>(null);
-  useClickOutside({ ref, ignoreRef: selectRef, open: isOpen, click: () => setIsOpen(false) });
+  useClickOutside({ ref, ignoreRef: selectRef, click: () => setIsOpen(false) });
 
   const toggleDropdown = () => setIsOpen(!isOpen);
   const handleOptionClick = (option: Option) => {
