+++ src/shared/ui/Input/SelectBox.tsx
@@ -19,7 +19,7 @@ export const SelectBox = ({ options, onChange, placeholder }: CustomSelectProps)
   const [selectedOption, setSelectedOption] = useState<Option | null>(null);
   const ref = useRef<HTMLUListElement>(null);
   const selectRef = useRef<HTMLDivElement>(null);
-  useClickOutside({ ref, ignoreRef: selectRef, open: isOpen, click: () => setIsOpen(false) });
+  useClickOutside({ ref, ignoreRef: selectRef, click: () => setIsOpen(false) });
 
   const toggleDropdown = () => setIsOpen(!isOpen);
   const handleOptionClick = (option: Option) => {
